Knarr now uses BV arithmetic to track variables

+++ knarr/src/main/java/edu/gmu/swe/knarr/runtime/PathUtils.java
@@ -216,7 +216,7 @@ public class PathUtils {
 
 	private static Expression getExpression(Taint<Expression> t, long v) {
 		if (t == null)
-			return new IntConstant((int) v);
+			return new IntConstant(v);
 		return t.lbl;
 	}
